Configuring OVM in separate truffle config as an easy example of side-by-side standard testing & ovm testing (#72)
