+++ packages/examples/package.json
@@ -4,11 +4,15 @@
   "description": "Example of using solc-transpiler & Web3 full node in place of solc-js and other Web3 providers",
   "scripts": {
     "clean": "rimraf build",
-    "build:truffle": "export EXECUTION_MANAGER_ADDRESS=\"0xA193E42526F1FEA8C99AF609dcEabf30C1c29fAA\" && truffle compile",
+    "build:truffle": "truffle compile",
+    "build:truffle-ovm": "truffle compile --config truffle-config-ovm.js",
     "build:waffle": "mkdir -p build/waffle && waffle waffle-config.json",
     "build": "yarn build:waffle && yarn build:truffle",
-    "test:truffle": "export EXECUTION_MANAGER_ADDRESS=\"0xA193E42526F1FEA8C99AF609dcEabf30C1c29fAA\" && truffle test --network fullnode ./truffle-tests/test-erc20.js",
-    "all": "yarn clean && yarn build"
+    "build:ovm": "mkdir -p build/waffle && waffle waffle-config.json && yarn build:truffle-ovm",
+    "test": "truffle test ./truffle-tests/test-erc20.js",
+    "test:ovm": "truffle test ./truffle-tests/test-erc20.js --config truffle-config-ovm.js",
+    "all": "yarn clean && yarn build && yarn test",
+    "all:ovm": "yarn clean && yarn build:ovm && yarn test:ovm"
   },
   "keywords": [
     "optimism",
