+++ packages/examples/truffle-tests/test-erc20.js
@@ -1,6 +1,5 @@
 const TestERC20 = artifacts.require("TestERC20");
 
-// TODO: Run `yarn server:fullnode` in rollup-full-node before executing this test
 contract('TestERC20', accounts => {
     const account = accounts[0];
     const account2 = accounts[1];
@@ -11,22 +10,22 @@ contract('TestERC20', accounts => {
     beforeEach(async () => {
         // Deploy a fresh contract every test
         erc20 = await TestERC20.new(initialValue, 'Test Token', 100, 'TEST',  {from: account});
-    })
+    });
 
     it('gets total supply', async () => {
         const totalSupply = await erc20.totalSupply.call();
         assert.equal(totalSupply, initialValue, 'Total supply mismatch!');
-    })
+    });
 
     it('gets balance of main account', async () => {
         const balance = await erc20.balanceOf.call(account);
         assert.equal(balance, initialValue, 'Sender balance mismatch!');
-    })
+    });
 
     it('gets 0 balance of second account', async () => {
         const balance = await erc20.balanceOf.call(account2);
         assert.equal(balance, 0, 'Second account balance mismatch!');
-    })
+    });
 
     it('transfers from account 1 to account 2', async () => {
         await erc20.transfer(account2, transferValue, {from: account});
@@ -36,7 +35,7 @@ contract('TestERC20', accounts => {
 
         const account2Balance = await erc20.balanceOf.call(account2);
         assert.equal(account2Balance, transferValue, 'Account balance mismatch after transfer!');
-    })
+    });
 
     it('transfers from account 1 to account 2 and then back', async () => {
         await erc20.transfer(account2, transferValue, {from: account});
@@ -47,5 +46,5 @@ contract('TestERC20', accounts => {
 
         const account2Balance = await erc20.balanceOf.call(account2);
         assert.equal(account2Balance, 0, 'Account balance mismatch after transfer!');
-    })
+    });
 });
